Changed nanos_create_wd() and nanos_create_wd_and_run() API functions. (Fixes #619)
'master' family has been changed to 5012.

The number of parameters has been reduced by grouping constant values in a new struct.

The name of these functions has been changed to:
nanos_create_wd_compact()
nanos_create_wd_and_run_compact()
**Note that old functions have been removed from the code.**

Also, the type of WD's data alignment has been changed from 'int' to 'size_t'.

+++ src/core/system.cpp
@@ -491,7 +491,7 @@ void System::finish ()
  *  +---------------+
  *
  */
-void System::createWD ( WD **uwd, size_t num_devices, nanos_device_t *devices, size_t data_size, int data_align,
+void System::createWD ( WD **uwd, size_t num_devices, nanos_device_t *devices, size_t data_size, size_t data_align,
                         void **data, WG *uwg, nanos_wd_props_t *props, size_t num_copies, nanos_copy_data_t **copies, nanos_translate_args_t translate_args )
 {
    ensure(num_devices > 0,"WorkDescriptor has no devices");
