+++ src/core/system_decl.hpp
@@ -170,7 +170,7 @@ namespace nanos
          void inlineWork ( WD &work );
 
          void createWD (WD **uwd, size_t num_devices, nanos_device_t *devices,
-                        size_t data_size, int data_align, void ** data, WG *uwg,
+                        size_t data_size, size_t data_align, void ** data, WG *uwg,
                         nanos_wd_props_t *props, size_t num_copies, nanos_copy_data_t **copies, nanos_translate_args_t translate_args );
 
          void createSlicedWD ( WD **uwd, size_t num_devices, nanos_device_t *devices, size_t outline_data_size,
